Optimize converter example
Turned on optimizations when compiling the converter. Also switched to doing the 16-bit conversion to an array, which then gets written to the output file in one go. This makes the converter more than three times as fast, making it competitive with xmp. This helps when doing batch conversion or testing.